requirements.txt
openai>=1.40.0
tiktoken>=0.7.0
python-dotenv>=1.0.1
rich>=13.7.1
colorama>=0.4.6
requests>=2.32.3
PyPDF2>=3.0.1
pyttsx3>=2.90
flask>=3.0.0
